pyproject.toml
[project]
name = "erp-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "amqp==5.3.1",
 "asgiref==3.8.1",
 "billiard==4.2.1",
 "celery==5.4.0",
 "certifi==2025.1.31",
 "charset-normalizer==3.4.1",
 "click==8.1.8",
 "click-didyoumean==0.3.1",
 "click-plugins==1.1.1",
 "click-repl==0.3.0",
 "django-filter>=25.1",
 "django==5.1.6",
 "django-extensions==3.2.3",
 "djangorestframework==3.15.2",
 "faker==36.1.1",
 "idna==3.10",
 "kombu==5.4.2",
 "markdown==3.7",
 "np==1.0.2",
 "numpy==2.2.3",
 "prompt-toolkit==3.0.50",
 "psycopg2-binary>=2.9.10",
 "python-dateutil==2.9.0.post0",
 "requests==2.32.3",
 "six==1.17.0",
 "sqlparse==0.5.3",
 "tqdm==4.67.1",
 "tzdata==2025.1",
 "urllib3==2.3.0",
 "vine==5.1.0",
 "wcwidth==0.2.13",
 "redis>=5.2.1",
 "jinja2>=3.1.6",
 "xhtml2pdf>=0.2.17",
]

[tool.ruff]
exclude = [
    "erp_system/settings/test.py"
]

[tool.mypy]
exclude = [
    "erp_system/settings/test.py"
]
